feat: Add exam and exam type management with repository, service, and controller support
Implemented CRUD operations for exams and exam types. Updated repositories with methods to handle academic year, school context, and assignment relations. Adjusted DTOs, models, and mappings for enhanced data consistency and functionality.

+++ src/main/java/com/edusyspro/api/dto/custom/AssignmentExhaustif.java
@@ -13,7 +13,7 @@ import java.time.ZonedDateTime;
 public record AssignmentExhaustif(
         Long id,
         Integer planningId,
-        Long examId,
+        Integer examId,
         ExamType examType,
         LocalDate examStart,
         LocalDate examEnd,
